parking_app/requirements.txt
streamlit
ultralytics
torch
torchvision
opencv-python-headless
Pillow
numpy